+++ forge/src/main/java/io/github/flemmli97/runecraftory/forge/RuneCraftoryForge.java
@@ -111,6 +111,7 @@ public class RuneCraftoryForge {
     public void registry(RegistryEvent.Register<GlobalLootModifierSerializer<?>> event) {
         ModLootCondition.LOOTFUNCTION.registerContent();
         ModLootCondition.LOOTCONDITIONS.registerContent();
+        ModLootCondition.LOOTNUMBERS.registerContent();
         ModStructures.STRUCTURESPROCESSORS.registerContent();
         ModCrafting.RECIPETYPE.registerContent();
     }