+++ forge/src/main/java/io/github/flemmli97/runecraftory/forge/platform/PlatformImpl.java
@@ -45,6 +45,7 @@ import net.minecraft.world.level.block.SoundType;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.storage.loot.LootContext;
 import net.minecraftforge.common.ForgeHooks;
 import net.minecraftforge.common.extensions.IForgeMenuType;
 import net.minecraftforge.event.ForgeEventFactory;
@@ -255,5 +256,10 @@ public class PlatformImpl implements Platform {
     @Override
     public float onLivingHurt(LivingEntity entity, DamageSource damageSrc, float damageAmount) {
         return ForgeHooks.onLivingHurt(entity, damageSrc, damageAmount);
+    }
+
+    @Override
+    public int getLootingFromCtx(LootContext ctx) {
+        return ctx.getLootingModifier();
     }
 }